backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
audioop-lts==0.2.1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
distro==1.9.0
executing==2.2.0
fastapi==0.115.7
google-api-core==2.24.0
google-auth==2.38.0
google-cloud-speech==2.30.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SpeechRecognition==3.14.1
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.45.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
